+++ sources/shared/shared/plugin/yPluginApi/StandardCapacities.cpp
@@ -25,6 +25,7 @@ namespace shared
          DECLARE_CAPACITY(CStandardCapacities::Direction, "direction", CStandardUnits::Degrees(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Distance, "distance", CStandardUnits::Meter(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Duration, "duration", CStandardUnits::NoUnit(), EKeywordDataType::kNumeric);
+         DECLARE_CAPACITY(CStandardCapacities::ElectricLoad, "electricLoad", CStandardUnits::AmperePerHOur(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Energy, "energy", CStandardUnits::WattPerHour(), EKeywordDataType::kNumeric);
          DECLARE_CAPACITY(CStandardCapacities::Event, "event", CStandardUnits::NoUnit(), EKeywordDataType::kNoData);
          DECLARE_CAPACITY(CStandardCapacities::Frequency, "frequency", CStandardUnits::Hertz(), EKeywordDataType::kNumeric);