+++ sources/shared/shared/plugin/yPluginApi/StandardUnits.h
@@ -19,6 +19,7 @@ namespace shared
             //--------------------------------------------------------------
             static const std::string& NoUnit();
             static const std::string& Ampere(); // A
+            static const std::string& AmperePerHOur(); // A
             static const std::string& CubicMetre(); // m3
             static const std::string& CubicMeterPerSecond(); // m3/s
             static const std::string& Decibel(); // dB