+++ sources/plugins/EnOcean/profiles/hardCoded/Profile_A5_12_05.h
@@ -1,0 +1,39 @@
+#pragma once
+#include "../IType.h"
+
+
+class CProfile_A5_12_05 : public IType
+{
+public:
+   CProfile_A5_12_05(const std::string& deviceId,
+                     boost::shared_ptr<yApi::IYPluginApi> api);
+   virtual ~CProfile_A5_12_05();
+
+   // IType implementation
+   const std::string& profile() const override;
+   const std::string& title() const override;
+   std::vector<boost::shared_ptr<const yApi::historization::IHistorizable>> allHistorizers() const override;
+   void readInitialState(const std::string& senderId,
+                         boost::shared_ptr<IMessageHandler> messageHandler) const override;
+   std::vector<boost::shared_ptr<const yApi::historization::IHistorizable>> states(unsigned char rorg,
+                                                                                   const boost::dynamic_bitset<>& data,
+                                                                                   const boost::dynamic_bitset<>& status,
+                                                                                   const std::string& senderId,
+                                                                                   boost::shared_ptr<IMessageHandler> messageHandler) const override;
+   void sendCommand(const std::string& keyword,
+                    const std::string& commandBody,
+                    const std::string& senderId,
+                    boost::shared_ptr<IMessageHandler> messageHandler) const override;
+   void sendConfiguration(const shared::CDataContainer& deviceConfiguration,
+                          const std::string& senderId,
+                          boost::shared_ptr<IMessageHandler> messageHandler) const override;
+   // [END] IType implementation
+
+private:
+   static const int NB_POSSESSED = 10;
+   const std::string m_deviceId;
+   boost::shared_ptr<yApi::historization::CSwitch> m_possessed[NB_POSSESSED];
+   boost::shared_ptr<yApi::historization::CTemperature> m_temperature;
+   boost::shared_ptr<yApi::historization::CBatteryLevel> m_battery;
+   std::vector<boost::shared_ptr<const yApi::historization::IHistorizable>> m_historizers;
+};