+++ sources/server/changelog.md
@@ -1,7 +1,9 @@
 ### 2.1.0-beta.2
 
 #### New features and fixes
-* Add debit historizer
+* Add debit historizer (m³/s)
+* Add electric load historizer (Ah)
+
 
 ### 2.1.0-beta.1